refactor(ts): misc routes and index
