--- src/service/routes/healthcheck.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = new express.Router();
-
-router.get('/', function (req, res) {
-  res.send({
-    message: 'ok',
-  });
-});
-
-module.exports = router;
