--- src/service/routes/home.js
@@ -1,14 +1,0 @@
-const express = require('express');
-const router = new express.Router();
-
-const resource = {
-  healthcheck: '/api/v1/healthcheck',
-  push: '/api/v1/push',
-  auth: '/api/auth',
-};
-
-router.get('/', function (req, res) {
-  res.send(resource);
-});
-
-module.exports = router;
